move calculate_gas_info_impl to runtime_api mod
